# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BestSub -  API 文档\n\n这是 BestSub 的 API 文档\n\n## 认证\n大多数接口需要使用 JWT 令牌进行认证。\n认证时，请在 Authorization 头中包含 JWT 令牌：\n`Authorization: Bearer \u003cyour-jwt-token\u003e`\n\n## 错误响应\n所有错误响应都遵循统一格式，包含 code、message 和 error 字段。\n\n## 成功响应\n所有成功响应都遵循统一格式，包含 code、message 和 data 字段。",
        "title": "BestSub API",
        "contact": {
            "name": "BestSub API 支持",
            "email": "support@bestsub.com"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://opensource.org/license/gpl-3-0"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/check": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "检测"
                ],
                "summary": "获取检测列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "检测ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_models_check.Response"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建单个检测",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "检测"
                ],
                "summary": "创建检测",
                "parameters": [
                    {
                        "description": "创建检测请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_models_check.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_models_check.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/v1/check/type": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取检测类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "检测"
                ],
                "summary": "获取检测类型",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_core_check.Desc"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/v1/check/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据请求体中的ID更新检测信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "检测"
                ],
                "summary": "更新检测",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "检测ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新检测请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_models_check.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_models_check.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "404": {
                        "description": "检测不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID删除单个检测",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "检测"
                ],
                "summary": "删除检测",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "检测ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "404": {
                        "description": "检测不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/v1/check/{id}/run": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "手动触发检测执行",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "检测"
                ],
                "summary": "手动运行检测",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "检测ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "运行成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "404": {
                        "description": "检测不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/v1/check/{id}/stop": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "停止正在运行的检测",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "检测"
                ],
                "summary": "停止检测",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "检测ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "停止成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "404": {
                        "description": "检测不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_bestruirui_bestsub_internal_core_check.Desc": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "string"
                },
                "require": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "github_com_bestruirui_bestsub_internal_models_check.Request": {
            "type": "object",
            "properties": {
                "config": {},
                "enable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "example": "测试检测任务"
                },
                "task": {
                    "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_models_check.Task"
                }
            }
        },
        "github_com_bestruirui_bestsub_internal_models_check.Response": {
            "type": "object",
            "properties": {
                "config": {},
                "enable": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_models_check.Result"
                },
                "status": {
                    "type": "string"
                },
                "task": {
                    "$ref": "#/definitions/github_com_bestruirui_bestsub_internal_models_check.Task"
                }
            }
        },
        "github_com_bestruirui_bestsub_internal_models_check.Result": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "extra": {},
                "last_run": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "github_com_bestruirui_bestsub_internal_models_check.Task": {
            "type": "object",
            "properties": {
                "cron_expr": {
                    "type": "string",
                    "example": "0 0 * * *"
                },
                "log_level": {
                    "type": "string",
                    "example": "info"
                },
                "log_write_file": {
                    "type": "boolean",
                    "example": true
                },
                "notify": {
                    "type": "boolean",
                    "example": true
                },
                "notify_channel": {
                    "type": "integer",
                    "example": 1
                },
                "sub_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1
                    ]
                },
                "sub_id_exclude": {
                    "type": "boolean",
                    "example": false
                },
                "timeout": {
                    "type": "integer",
                    "example": 60
                },
                "type": {
                    "type": "string",
                    "example": "test"
                }
            }
        },
...
        "github_com_bestruirui_bestsub_internal_server_resp.ResponseStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "类型为 \"Bearer\"，后跟空格和 JWT 令牌。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}